Initialize config and ctrl_meas structs to power-on-reset state, issue #60
Initialize Adafruit_BMP280::_config and _ctrlMeas members to power-on-reset state in their constructors. Note that Adafruit_BMP280::begin() calls Adafruit_BMP280::setSampling() which will mostly initialize these structs as well. The defect in issue #60 is that _config.spi3w_en is never initialized and consequently the sensor can end up configured for spi 3-wire mode, which is not supported by the code. Fix is to initialize all members of these configuration structs to the power-on-reset values defined in the datasheet.